Config redis package

+++ rental_app/app/Module/File/Type/Price.php
@@ -14,7 +14,7 @@ final class Price
         $this->value = bcdiv($value, '1', 2);
     }
 
-    public function getDataBaseValue(): int
+    public function getDataBaseValue(): float
     {
         return $this->value * 100;
     }
